+++ src/actions/wallet/signTransaction.test.ts
@@ -401,36 +401,6 @@ describe('custom (cip64)', () => {
   })
 })
 
-describe('custom (cip42)', () => {
-  const client = createWalletClient({
-    chain: celo,
-    transport: http(anvilMainnet.rpcUrl.http),
-  })
-  const tx = {
-    account: privateKeyToAccount(sourceAccount.privateKey),
-    chain: null,
-    ...base,
-    feeCurrency: '0x765de816845861e75a25fca122bb6898b8b1282a',
-    maxFeePerGas: parseGwei('20'),
-    maxPriorityFeePerGas: parseGwei('2'),
-    gatewayFee: 4n,
-    gatewayFeeRecipient: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
-  } as const
-
-  test('default', async () => {
-    expect(await signTransaction(client, tx)).toMatchInlineSnapshot(
-      '"0x7cf8840182031184773594008504a817c80082520894765de816845861e75a25fca122bb6898b8b1282a940f16e9b0d03470827a95cdfd0cb8a8a3b46969b904808080c001a062bee7f81cccd1f430b4b66ec5a23737d6fbee9965e63ac582d09f63aef32bdca05e75bd3ef63f2c0f6fd0a87e3f8d4809a38ac955b7d97fd4af1bd2c882999d5c"',
-    )
-  })
-  test('sanity', async () => {
-    expect(
-      await signTransaction(client, { ...tx, type: 'cip42' }),
-    ).toMatchInlineSnapshot(
-      '"0x7cf8840182031184773594008504a817c80082520894765de816845861e75a25fca122bb6898b8b1282a940f16e9b0d03470827a95cdfd0cb8a8a3b46969b904808080c001a062bee7f81cccd1f430b4b66ec5a23737d6fbee9965e63ac582d09f63aef32bdca05e75bd3ef63f2c0f6fd0a87e3f8d4809a38ac955b7d97fd4af1bd2c882999d5c"',
-    )
-  })
-})
-
 describe('errors', () => {
   test('no account', async () => {
     await expect(() =>