+++ src/celo/utils.ts
@@ -3,7 +3,6 @@ import { trim } from '../utils/data/trim.js'
 import type {
   CeloTransactionRequest,
   CeloTransactionSerializable,
-  TransactionSerializableCIP42,
   TransactionSerializableCIP64,
 } from './types.js'
 
@@ -38,23 +37,6 @@ export function isEIP1559(
   )
 }
 
-// process as CIP42 if any of these fields are present. realistically gatewayfee is not used but is part of spec
-export function isCIP42(
-  transaction: CeloTransactionSerializable | CeloTransactionRequest,
-): transaction is TransactionSerializableCIP42 {
-  // Enable end-user to force the tx to be considered as a cip42
-  if (transaction.type === 'cip42') {
-    return true
-  }
-
-  return (
-    isEIP1559(transaction) &&
-    (isPresent(transaction.feeCurrency) ||
-      isPresent(transaction.gatewayFeeRecipient) ||
-      isPresent(transaction.gatewayFee))
-  )
-}
-
 export function isCIP64(
   transaction: CeloTransactionSerializable | CeloTransactionRequest,
 ): transaction is TransactionSerializableCIP64 {
@@ -72,10 +54,5 @@ export function isCIP64(
     return true
   }
 
-  return (
-    isEIP1559(transaction) &&
-    isPresent(transaction.feeCurrency) &&
-    isEmpty(transaction.gatewayFee) &&
-    isEmpty(transaction.gatewayFeeRecipient)
-  )
+  return isEIP1559(transaction) && isPresent(transaction.feeCurrency)
 }