+++ src/zksync/actions/estimateGasL1ToL2.test.ts
@@ -1,0 +1,31 @@
+import { expect, test } from 'vitest'
+import {
+  mockClientPublicActionsL2,
+  zkSyncClientLocalNode,
+  zkSyncClientLocalNodeWithAccount,
+} from '../../../test/src/zksync.js'
+import { estimateGasL1ToL2 } from './estimateGasL1ToL2.js'
+
+const client = { ...zkSyncClientLocalNode }
+
+const clientWithAccount = { ...zkSyncClientLocalNodeWithAccount }
+
+mockClientPublicActionsL2(client)
+mockClientPublicActionsL2(clientWithAccount)
+
+test('default with account', async () => {
+  const value = await estimateGasL1ToL2(clientWithAccount, {
+    to: '0xa61464658AfeAf65CccaaFD3a512b69A83B77618',
+    value: 70n,
+  })
+  expect(value > 0).to.be.true
+})
+
+test('default without account', async () => {
+  const value = await estimateGasL1ToL2(client, {
+    account: '0x36615Cf349d7F6344891B1e7CA7C72883F5dc049',
+    to: '0xa61464658AfeAf65CccaaFD3a512b69A83B77618',
+    value: 70n,
+  })
+  expect(value > 0).to.be.true
+})